Keep mqtt config details in its own file
